python-p2p/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiofiles==23.2.1
python-multipart==0.0.6
pyyaml==6.0.1